Fix build nav bug
