+++ src/components/Sidebar/SidebarItem/NonExpandableHeading.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ROOT_LEVEL } from '../consts';
-import { AccordionHeadingProps } from './accordion-heading-props';
+import { AccordionHeadingProps } from './ExpandableAccordionHeading';
 
 export const NonExpandableHeading = ({ label, level }: AccordionHeadingProps) => (
   <div aria-level={level || ROOT_LEVEL}>{label}</div>
