--- src/components/Sidebar/SidebarItem/accordion-heading-props.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export type AccordionHeadingProps = {
-  label: React.ReactNode;
-  level: number;
-  expandable?: boolean;
-};
