+++ src/components/Sidebar/SidebarItem/index.ts
@@ -1,1 +1,2 @@
 export * from './SidebarItem';
+export { AccordionHeadingProps } from './ExpandableAccordionHeading';
