+++ src/contexts/index.ts
@@ -1,3 +1,4 @@
 export * from './page-language-context';
 export { default as PageLanguageContext } from './page-language-context';
 export * from './user-context';
+export * from './pathname-context';
