+++ src/contexts/pathname-context.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const PathnameContext = createContext<string>('');
